+++ examples/exports/run.cjs.js
@@ -1,0 +1,3 @@
+const assert = require("assert");
+const { run } = require("cypress-cloud");
+assert(typeof run === "function");