+++ packages/cypress-cloud/lib/stdout.ts
@@ -7,7 +7,6 @@ cp.spawn = function (command, args, options) {
   // @ts-ignore
   if (command.match(/Cypress/)) {
     // @ts-ignore
-    debug;
     const process = orginal(command, args, {
       ...options,
       // using pipe enables capturing stdout and stderr