+++ packages/cypress-cloud/publish.js
@@ -22,10 +22,17 @@ if (!options.tag) {
 const newPkg = {
   ...pkg,
   main: "./index.js",
+  bin: {
+    "cypress-cloud": "./bin/cli.js",
+  },
   files: ["*"],
   bin: "./bin/cli.js",
   exports: {
-    ".": "./",
+    ".": {
+      import: "./index.mjs",
+      require: "./index.js",
+      types: "./index.d.ts",
+    },
     "./plugin": {
       import: "./plugin/index.js",
       require: "./plugin/index.js",